Fixed an obscure bug in which certificates may not be found.
Not sure if anyone else has run into this issue, but I did in a separate project.
The issue is that when the signature node has been serialized to a String,
it loses the context of where it is in the XML tree. When deserialized, it
only knows about itself. Later in the code, in valdiateSignatureValue of xml-crypto,
this is calleD:

However, if the signature node has been serialized and deserialized, it
incorrectly reports it only has 1 child (itself), as opposed to the correct number.
This causes the certificate data to not be found, and the validation to fail.

All tests still pass with this change. Update to xml-crypto required.