Support multiple and dynamic signing certificates (#218)
Some organizations roll the signing certificate every year.  During the roll over period the server provides two certificates which may be valid, so when a SAML response is provided it is valid if it is signed with either of those signatures.

This change allows the `cert` to ban an array or a async function which returns a single cert or an array of certs.