Multiple callback urls
Multiple callback urls can be now be defined and will be generated within the metadata.xml
Only the default callback is used when sending AuthnRequests. If no default is set, the first option is used as the default.

Can either be set as an array of objects or an array of strings. Existing implementations can still just assign a singular string.