internal: provide unique failure messages for invalid signatures. Fixes #146
Before, there were 6 different cases in which "Invalid Signature" could
be returned, making it hard to debug which case had been triggered.

Now we return a unique message in each case. For improvement backwards
compatibility, all cases still contain the string "Invalid signature",
so code that was checking for that regex will be fine.

If code was checking for the exact string "Invalid Signature", it will
need to be updated.